requirements.txt
sentence-transformers
chromadb
python-docx
pypdf
reportlab
openpyxl
xlrd
